+++ repositories/OrderProductRepository.java
@@ -3,5 +3,5 @@ package com.Bootcamp.Project.Application.repositories;
 import com.Bootcamp.Project.Application.entities.OrderProduct;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderProductRepository extends CrudRepository<OrderProduct,Integer> {
+public interface OrderProductRepository extends CrudRepository<OrderProduct,Long> {
 }
